Fixed collision code + Added Basic UI
Now energy is transferred correctly for collisions.

UI displays a timer going down until the end of a wave, It also displays the target score for the wave, and current score. No logic is associated with any of them yet.

Timer object created which can be used globally. This should be used for mission timers and also for tracking how much time remains in the current wave.