Added Escort Mission Logic For Dead AI
Added new logic. If AI dies, we award points proportional to number of AI that made it. If no AI made it, mission autocompletes with 0 points awarded. If all make it, get full score.

Edited the Planet flair code so that if a planet holds multiple missions then the pick-up flair holds until